+++ src/correlation.h
@@ -2,8 +2,8 @@
 #define CORRELATION_H
 
 int read_data_file(char *filename, struct Model * m);
-int write_correlation_function_emf(char * fn, double T, double dT, double taus, double S2s, double tauf, double S2f);
-int write_correlation_function_smf(char * fn, double T, double dT, double tau, double S2);
+int write_correlation_function_emf(char * fn, Decimal T, Decimal dT, Decimal taus, Decimal S2s, Decimal tauf, Decimal S2f);
+int write_correlation_function_smf(char * fn, Decimal T, Decimal dT, Decimal tau, Decimal S2);
 int main(int argc, char * argv[]);
 
 #include "correlation.c"
