+++ src/crosen.h
@@ -1,7 +1,7 @@
 #ifndef CROSEN_H
 #define CROSEN_H
 
-double simplex(double (*func)(double[], struct Residue*, struct Model *, unsigned int), double start[], double EPSILON, double scale, struct Residue * resid, struct Model * mod);
+Decimal simplex(Decimal (*func)(Decimal[], struct Residue*, struct Model *, unsigned int), Decimal start[], Decimal EPSILON, Decimal scale, struct Residue * resid, struct Model * mod);
 
 #include "crosen.c"
 #endif
